fix & move LoadAverage sample widget to samples/widgets
