+++ samples/widgets/LoadAverage/package.json
