+++ samples/widgets/LoadAverage/src/Widget.jsx
