+++ samples/widgets/LoadAverage/webpack.config.js
@@ -1,13 +1,14 @@
 const path = require('path');
 const webpack = require('webpack')
 module.exports = {
+    devtool: 'source-map',
     context: path.resolve(__dirname, './src'),
     entry: {
         index: './LoadAverage.jsx',
     },
     output: {
         path: path.resolve(__dirname, './dist/'),
-        filename: 'bundle.js',
+        filename: 'LoadAverage.js',
     },
     module: {
         loaders: [
